Add test client connecting module
